M1809_GetData功能正常
