+++ easydo/Algorithm/M1809/output/compare_industry.csv
@@ -1,0 +1,4 @@
+,总资产,净资产,资产负债比,流动资产,一年内到期的长期负债,应收账款,预收账款,存货,营业收入,营业成本,营业税金及附加,财务费用,营业外收入,净利润,除非净利润,每股收益,经营净额,投资净额,筹资净额,汇率影响,现金净增加额,期末现金余额,流动比率,资产周转率,存货周转率,溢价比,市盈率,市净率,名义净资产收益率,真实净资产收益率,毛利率,营收增长率,除非净利润增长率,股息率,分红率
+000651,21496799.93,6683479.77,0.69,17153464.62,0.0,581449.16,1414303.82,1656834.72,15001955.16,12469881.27,151303.54,43128.47,51105.91,2250859.9,2188521.9,3.72,1635853.82,-6225345.8,-226934.12,-179802.74,4996228.83,7132136.0,1.16,1.18,15.05,3.96,11.75,11.05,4.75,0.43,0.2,0.36,-2.13,0.045,0.484
+000333,24810685.8,8292517.1000000015,0.67,16981067.6,13660.5,1752871.7,1740906.3,2944416.6,24191889.6,22473459.2,141642.8,81594.9,46720.4,1861119.0,1682316.1,2.66,2444262.3,-3473960.8,1965164.5,-3673.7,-931792.2999999998,1251373.0,1.43,1.62,15.27,4.68,20.84,11.85,3.67,0.31,0.08,0.51,-2.23,0.028,0.376
+avarage,23153742.865000002,7487998.4350000005,0.6799999999999999,17067266.11,6830.25,1167160.43,1577605.06,2300625.66,19596922.380000003,17471670.235,146473.16999999998,62361.685,48913.155,2055989.45,1935419.0,3.1900000000000004,2040058.06,-4849653.3,869115.19,-91738.22,2032218.2650000001,4191754.5,1.295,1.4,15.16,4.32,16.295,11.45,4.21,0.37,0.14,0.435,-2.1799999999999997,0.0365,0.43
