+++ easydo/Algorithm/M1809/output/compare_self.csv
@@ -1,0 +1,9 @@
+,总资产,净资产,资产负债比,流动资产,一年内到期的长期负债,应收账款,预收账款,存货,营业收入,营业成本,营业税金及附加,财务费用,营业外收入,净利润,除非净利润,每股收益,经营净额,投资净额,筹资净额,汇率影响,现金净增加额,期末现金余额,流动比率,资产周转率,存货周转率,溢价比,市盈率,市净率,名义净资产收益率,真实净资产收益率,毛利率,营收增长率,除非净利润增长率,股息率,分红率
+2010,6560437.81,1401170.8199999994,0.79,5453271.86,0.0,119892.49,1200621.09,1155916.73,6080724.25,5819180.99,53881.54,-30896.95,235332.41,430320.55,203106.37,1.52,61599.5,-188705.99,-141519.87,-7659.19,276285.54000000004,947337.05,1.1,1.63,9.77,3.66,11.93,4.95,1.88,0.38,0.04,0.43,-1.92,0.024,0.329
+2011,8521159.42,1837715.4500000002,0.78,7175561.05,217661.57,122679.38,1975269.37,1750310.73,8351725.25,7900847.55,49790.8,-45270.76,184578.56,529734.05,339939.77,1.86,335616.0,-276727.25,-79460.07,-16597.41,37168.72999999998,671051.51,1.12,1.73,8.38,2.68,9.3,6.45,2.45,0.38,0.06,0.37,-184.73,0.014,0.161
+2012,10756689.99,2758020.21,0.74,8508764.51,251376.37,147487.3,1663011.32,1723504.26,10011010.88,9231019.15,58995.36,-46134.76,76038.32,744592.8,679555.81,2.47,1840874.63,-421252.4,81521.65,2022.02,-1503165.89,633882.78,1.08,1.65,10.71,2.79,10.32,9.15,3.57,0.39,0.08,0.2,-4.3,0.023,0.202
+2013,13370210.34,3546667.7699999996,0.73,10373252.22,92344.8,184927.53,1198643.37,1312273.04,12004307.0,10948792.61,95616.98,-13730.86,68419.96,1093575.52,978189.72,3.61,1296983.71,-218598.74,-242397.72,-47117.62,-788869.6400000001,2137048.67,1.08,1.54,16.69,2.79,9.05,11.71,4.57,0.39,0.1,0.2,-2.47,0.039,0.277
+2014,15623094.85,4513145.1,0.71,12014347.88,206149.09,266134.76,642772.24,859909.81,14000539.4,12325897.96,136242.49,-94224.47,70606.38,1425295.48,1326380.42,4.71,1893916.55,-286213.79,-186431.39,3457.43,-1424728.8000000003,2925918.31,1.11,1.5,28.67,2.49,7.88,14.91,5.81,0.39,0.14,0.17,-2.23,0.048,0.318
+2015,16169801.63,4856660.860000001,0.7,12094931.46,240374.56,287921.21,761959.8,947394.27,10056445.36,8613460.91,75189.42,-192879.73,140429.17,1262373.26,1122015.32,2.08,4437838.18,-471315.49,-768302.2,187634.08,-3385854.5699999994,4350647.11,1.07,1.02,18.18,2.86,10.99,8.0,2.48,0.31,0.17,-0.28,-1.9,0.051,1.442
+2016,18236970.5,5492360.27,0.7,14291078.35,0.0,296053.47,1002188.55,902490.52,11011310.19,9152937.98,143040.42,-484554.66,109623.48,1552463.49,1666950.73,2.56,1485995.21,-1924655.27,-575156.02,409450.4,604365.6799999997,7736501.68,1.13,0.99,18.26,2.72,9.62,9.06,2.99,0.33,0.2,0.09,-2.55,-1.0,0.586
+2017,21496799.93,6683479.77,0.69,17153464.62,0.0,581449.16,1414303.82,1656834.72,15001955.16,12469881.27,151303.54,43128.47,51105.91,2250859.9,2188521.9,3.72,1635853.82,-6225345.8,-226934.12,-179802.74,4996228.83,7132136.0,1.16,1.18,15.05,3.96,11.75,11.05,4.75,0.43,0.2,0.36,-2.13,0.045,0.484
