+++ easydo/Algorithm/M1809/output/normalize.csv
@@ -1,0 +1,4 @@
+,总资产,净资产,资产负债比,流动资产,一年内到期的长期负债,应收账款,预收账款,存货,营业收入,营业成本,营业税金及附加,财务费用,营业外收入,净利润,除非净利润,每股收益,经营净额,投资净额,筹资净额,汇率影响,现金净增加额,期末现金余额,流动比率,资产周转率,存货周转率,溢价比,市盈率,市净率,名义净资产收益率,真实净资产收益率,毛利率,营收增长率,除非净利润增长率,股息率,分红率
+000651,21496799.93,6683479.77,0.69,0.798,0.0,0.027,0.066,0.077,0.698,0.831,0.01,0.003,0.003,0.15,0.146,0.146,0.109,-0.415,-0.015,-0.012,0.333,0.475,1.16,1.18,15.05,3.96,11.75,11.05,4.75,0.43,0.2,0.36,-2.13,0.045,0.484
+000333,24810685.8,8292517.1000000015,0.67,0.684,0.001,0.071,0.07,0.119,0.975,0.929,0.006,0.003,0.002,0.077,0.07,0.07,0.101,-0.144,0.081,-0.0,-0.039,0.052,1.43,1.62,15.27,4.68,20.84,11.85,3.67,0.31,0.08,0.51,-2.23,0.028,0.376
+avarage,23153742.865000002,7487998.4350000005,0.6799999999999999,0.737,0.0,0.05,0.068,0.099,0.846,0.892,0.007,0.003,0.002,0.105,0.099,0.099,0.104,-0.247,0.044,-0.005,0.104,0.214,1.295,1.4,15.16,4.32,16.295,11.45,4.21,0.37,0.14,0.435,-2.1799999999999997,0.0365,0.43
